Make a playerReducer to handle all toggle form, add and remove player
