+++ src/redux/actions/removePlayer.js
@@ -1,8 +1,8 @@
 import { REMOVE_PLAYER } from '../constants';
 
-const removePlayer = (player) => ({
+const removePlayer = (id) => ({
   type: REMOVE_PLAYER,
-  payload: player.id,
+  payload: { id },
 });
 
 export default removePlayer;
