+++ src/redux/constants.js
@@ -1,5 +1,8 @@
 const SHOW_FORM = 'SHOW_FORM';
 const ADD_PLAYER = 'ADD_PLAYER';
 const REMOVE_PLAYER = 'REMOVE_PLAYER';
+const initialState = { players: [], toggle: false };
 
-export { SHOW_FORM, ADD_PLAYER, REMOVE_PLAYER };
+export {
+  SHOW_FORM, ADD_PLAYER, REMOVE_PLAYER, initialState,
+};
