--- src/redux/reducers/showFormReducer.js
@@ -1,13 +1,0 @@
-import { SHOW_FORM } from '../constants';
-
-const initialState = { toggle: false };
-const showFormReducer = (state = initialState, action) => {
-  switch (action.type) {
-    case SHOW_FORM:
-      return { ...state, toggle: !state.toggle };
-    default:
-      return { ...state };
-  }
-};
-
-export default showFormReducer;
